chore(cubeorchestrator): share bench helpers via benches/common
